requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
arxiv==2.2.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.46.0
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
cassandra-driver==3.29.2
cassio==0.1.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
cryptography==45.0.6
dataclasses-json==0.6.7
ddgs==9.6.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.9
emoji==2.14.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.11
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2025.3.0
geomet==0.2.1.post1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.23.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.2
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kubernetes==32.0.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.66
langchain-groq==0.3.2
langchain-objectbox==0.1.0
langchain-ollama==0.2.3
langchain-openai==0.3.8
langchain-text-splitters==0.3.8
langdetect==1.0.9
langserve==0.3.1
langsmith==0.3.45
libmagic==1.0
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
objectbox==4.0.0
olefile==0.47
ollama==0.4.7
onnxruntime==1.21.0
openai==1.84.0
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==11.1.0
pip-system-certs==4.0
platformdirs==4.3.6
pluggy==1.6.0
posthog==3.21.0
primp==0.15.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.39
sqlite-fts4==1.0.3
sqlite-utils==3.38
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.43.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.18.13
unstructured-client==0.42.3
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
